Refactor gov e2e tests
